Support HA using queue and exchange federation.
This commit implements 2 changes that allow for the use of Rabbit's
federation plugin to implement HA.

a) Uses a different exchange for the reply queues. This allows to
federate the whole reply exchange so replies are copied to all
federated nodes and a client will get its response even if it's
waiting on a different node than the server.

b) Uses the original routing key for messages forwarded to another
queue through the federation mechanism.

HA can be implemented by federating the whole reply exchange and just
the RPC queue from the nameko-rpc exchange.